backend/requirements.txt

fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
celery
redis
httpx
pytest
pytest-asyncio
textstat
python-multipart